+++ src/Commands/SystemCommands/GenericmessageCommand.php
@@ -68,9 +68,57 @@ class GenericmessageCommand extends SystemCommand
 
         //Fetch conversation command if it exists and execute it
         if ($conversation->exists() && ($command = $conversation->getCommand())) {
-            return $this->telegram->executeCommand($command);
+            return $this->getTelegram()->executeCommand($command);
+        }
+
+        // Try to execute any deprecated system commands.
+        if ($deprecated_system_command_response = $this->executeDeprecatedSystemCommand()) {
+            return $deprecated_system_command_response;
         }
 
         return Request::emptyResponse();
+    }
+
+
+    public function executeDeprecatedSystemCommand()
+    {
+        $telegram = $this->getTelegram();
+        $update   = $this->getUpdate();
+        $message  = $this->getMessage();
+
+        // List of service messages previously handled internally.
+        $service_messages = [
+            'editedmessage'         => [$update, 'getEditedMessage'],
+            'channelpost'           => [$update, 'getChannelPost'],
+            'editedchannelpost'     => [$update, 'getEditedChannelPost'],
+            'newchatmembers'        => [$message, 'getNewChatMembers'],
+            'leftchatmember'        => [$message, 'getLeftChatMember'],
+            'newchattitle'          => [$message, 'getNewChatTitle'],
+            'newchatphoto'          => [$message, 'getNewChatPhoto'],
+            'deletechatphoto'       => [$message, 'getDeleteChatPhoto'],
+            'groupchatcreated'      => [$message, 'getGroupChatCreated'],
+            'supergroupchatcreated' => [$message, 'getSupergroupChatCreated'],
+            'channelchatcreated'    => [$message, 'getChannelChatCreated'],
+            'migratefromchatid'     => [$message, 'getMigrateFromChatId'],
+            'migratetochatid'       => [$message, 'getMigrateToChatId'],
+            'pinnedmessage'         => [$message, 'getPinnedMessage'],
+            'successfulpayment'     => [$message, 'getSuccessfulPayment'],
+        ];
+
+        foreach ($service_messages as $command => $service_message_getter) {
+            // Let's check if this message is a service message.
+            if ($service_message_getter() === null) {
+                continue;
+            }
+
+            // Make sure the command exists, otherwise GenericCommand would be executed instead!
+            if ($telegram->getCommandObject($command) === null) {
+                break;
+            }
+
+            return $telegram->executeCommand($command);
+        }
+
+        return null;
     }
 }